+++ Assets/00_DodgeBall/Scripts/Energy/IEnergyAction.cs
@@ -1,11 +1,8 @@
 ﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
 
 public interface IEnergyAction
 {
-    Action<float> ConsumeEnergy { get; set; }
-    Func<float,bool> CanConsumeEnergy { get; set; }
+    Action<int> ConsumeEnergy { get; set; }
+    Func<int,bool> CanConsumeEnergy { get; set; }
     Func<bool> AllowRegen { get; }
 }
